fix dart2js source map source uris and collapsing of sound mode dart sdk lines
