+++ build_web_compilers/lib/src/dev_compiler_builder.dart
@@ -320,25 +320,6 @@ String _sourceArg(AssetId id) {
   return uri.startsWith('package:') ? uri : '$multiRootScheme:///${id.path}';
 }
 
-/// Copied to `web/stack_trace_mapper.dart`, these need to be kept in sync.
-///
-/// Given a list of [uris] as [String]s from a sourcemap, fixes them up so that
-/// they make sense in a browser context.
-///
-/// - Strips the scheme from the uri
-/// - Strips the top level directory if its not `packages`
-List<String> fixSourceMapSources(List<String> uris) {
-  return uris.map((source) {
-    var uri = Uri.parse(source);
-    // We only want to rewrite multi-root scheme uris.
-    if (uri.scheme.isEmpty) return source;
-    var newSegments = uri.pathSegments.first == 'packages'
-        ? uri.pathSegments
-        : uri.pathSegments.skip(1);
-    return Uri(path: p.url.joinAll(['/'].followedBy(newSegments))).toString();
-  }).toList();
-}
-
 /// The module name according to ddc for [jsId] which represents the real js
 /// module file.
 String ddcModuleName(AssetId jsId, bool soundNullSafety) {
