+++ build_web_compilers/web/source_map_stack_trace.dart
@@ -62,8 +62,8 @@ StackTrace mapStackTrace(Mapping sourceMap, StackTrace stackTrace,
     }
 
     if (!sourceUrl.startsWith('dart:') &&
-        sourceUrl ==
-            'package:build_web_compilers/src/dev_compiler/dart_sdk.js') {
+        sourceUrl.startsWith(
+            'package:build_web_compilers/src/dev_compiler/dart_sdk.')) {
       // This compresses the long dart_sdk URLs if SDK source maps are missing.
       // It's no longer linkable, but neither are the properly mapped ones
       // above.
